requirements.txt
flask
flask-cors
gunicorn
google-generativeai